pyproject.toml
[tool.poetry]
name = "projetolumiboxback"
version = "0.1.0"
description = ""
authors = ["LuckeG <lucas.gabriel01@aluno.ifce.edu.br>"]
readme = "README.md"
packages = [{include = "projetolumiboxback", from = "src"}]


[tool.poetry.dependencies]
python = ">=3.10, <4.0"
django = ">=5.2,<6.0"
python-dotenv = ">=1.1.0,<2.0.0"
djangorestframework = "^3.16.0"
django-filter = "^25.1"
djoser = "^2.3.1"
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.7.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pylint = "^3.3.7"
pylint-django = "^2.6.1"
taskipy = "^1.14.1"
pre-commit = "^4.2.0"

[tool.taskipy.tasks]
qos = { cmd = "black src & isort src & pylint src", help = "Executa todas as ferramentas de qualidade de código."}
dev = { cmd = "python src/manage.py runserver"}
mak = { cmd = "python src/manage.py makemigrations api_lb"}
mig = { cmd = "python src/manage.py migrate"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
